context-processor/pyproject.toml
[project]
name = "crosswind-context"
version = "0.1.0"
description = "Document text extraction for Crosswind security evaluation platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "motor>=3.3.0",
    "docling>=2.0.0",
    "openpyxl>=3.1.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
# GCS support (optional, for users who want cloud storage)
gcs = ["google-cloud-storage>=2.10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crosswind_context"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
